+++ R/apportion_prior_healthcare.R
@@ -104,7 +104,7 @@ apportion_prior_healthcare <- function(patient_location, patient_category,
   }
 
   if(!requireNamespace("assertthat", quietly = TRUE)) {
-    stop("ggplot2 needed for this function to work. Please install it.",
+    stop("assertthat needed for this function to work. Please install it.",
          call. = FALSE)
   }
 
@@ -119,8 +119,8 @@ apportion_prior_healthcare <- function(patient_location, patient_category,
                           msg = "date_record_created must be a date")
 
   assertthat::assert_that(
-    length(adm_3_mo[!(tolower(adm_3_mo) %in% c("yes", "no", "don't know", NA))]) == 0,
-    msg = "adm_3_mo must be one of yes, no, don't know or NA")
+    length(adm_3_mo[!(tolower(adm_3_mo) %in% c("yes", "no", "don't know", "", NA))]) == 0,
+    msg = "adm_3_mo must be one of yes, no, don't know, \"\" or NA")
 
 
 
@@ -128,6 +128,10 @@ apportion_prior_healthcare <- function(patient_location, patient_category,
   adm_4_weeks <- tolower(adm_4_weeks)
   adm_12_weeks <- tolower(adm_12_weeks)
   adm_3_mo <- tolower(adm_3_mo)
+
+  adm_4_weeks <- ifelse(adm_4_weeks == "", NA_character_, adm_4_weeks)
+  adm_12_weeks <- ifelse(adm_12_weeks == "", NA_character_, adm_12_weeks)
+  adm_3_mo <- ifelse(adm_3_mo == "", NA_character_, adm_3_mo)
 
   hoha <- is_hoha(patient_location, patient_category, adm_date, spec_date, days_diff,
                   date_record_created)
@@ -140,15 +144,47 @@ apportion_prior_healthcare <- function(patient_location, patient_category,
                      ifelse(hoha == 0 & coha == 0 & coia == 1, "coia",
                             ifelse(hoha == 0 & coha == 0 & coia == 0 & coca == 1, "coca", NA)
                      )))
+  # z <- ifelse(adm_3_mo == "don't know" & !is.na(adm_3_mo) & z != "hoha", "unknown_3_mo", z)
+  # z <- ifelse((adm_3_mo == "no"  & !is.na(adm_3_mo)) &
+  #               (adm_4_weeks == "no"  & !is.na(adm_4_weeks)) &
+  #               (adm_12_weeks == "yes" & !is.na(adm_4_weeks)), "coia", z)
+  # z <- ifelse((adm_3_mo == "no"  & !is.na(adm_3_mo)) &
+  #               (adm_4_weeks == "yes"  & !is.na(adm_4_weeks)) &
+  #               (adm_12_weeks == "yes" & !is.na(adm_12_weeks)), "coha", z)
+  # z <- ifelse(z != "hoha" & is.na(adm_3_mo) & is.na(adm_4_weeks) &
+  #               is.na(adm_12_weeks), "all_blank", z)
+
   z <- ifelse(adm_3_mo == "don't know" & !is.na(adm_3_mo) & z != "hoha", "unknown_3_mo", z)
-  z <- ifelse((adm_3_mo == "no"  & !is.na(adm_3_mo)) &
-                (adm_4_weeks == "no"  & !is.na(adm_4_weeks)) &
-                (adm_12_weeks == "yes" & !is.na(adm_4_weeks)), "coia", z)
-  z <- ifelse((adm_3_mo == "no"  & !is.na(adm_3_mo)) &
-                (adm_4_weeks == "yes"  & !is.na(adm_4_weeks)) &
-                (adm_12_weeks == "yes" & !is.na(adm_12_weeks)), "coha", z)
-  z <- ifelse(z != "hoha" & is.na(adm_3_mo) & is.na(adm_4_weeks) &
-                is.na(adm_12_weeks), "all_blank", z)
+  z <- ifelse(adm_3_mo == "no" & !is.na(adm_3_mo) & z != "hoha", "coca", z)
+  z <- ifelse(is.na(adm_3_mo) & z != "hoha" &
+                !(is.na(adm_3_mo) & is.na(adm_4_weeks) & is.na(adm_12_weeks)),
+              "coca", z)
+  z <- ifelse(z != "hoha" & (is.na(adm_3_mo) & is.na(adm_4_weeks) & is.na(adm_12_weeks)),
+                             "all_blank", z)
+  # z <- ifelse((adm_3_mo == "no"  & !is.na(adm_3_mo)) &
+  #               (adm_4_weeks == "no"  & !is.na(adm_4_weeks)) &
+  #               (adm_12_weeks == "yes" & !is.na(adm_4_weeks)), "coca", z)
+  # z <- ifelse((adm_3_mo == "no"  & !is.na(adm_3_mo)) &
+  #               (adm_4_weeks == "yes"  & !is.na(adm_4_weeks)) &
+  #               (adm_12_weeks == "no" & !is.na(adm_4_weeks) | is.na(adm_4_weeks)), "coca", z)
+  # z <- ifelse((adm_3_mo == "no"  & !is.na(adm_3_mo)) &
+  #               (adm_4_weeks == "yes"  & !is.na(adm_4_weeks)) &
+  #               (adm_12_weeks == "yes" & !is.na(adm_12_weeks)), "coca", z)
+  #
+  # z <- ifelse(is.na(adm_3_mo) &
+  #               (adm_4_weeks == "no"  & !is.na(adm_4_weeks)) &
+  #               (adm_12_weeks == "yes" & !is.na(adm_4_weeks)), "coca", z)
+  # z <- ifelse(is.na(adm_3_mo) &
+  #               (adm_4_weeks == "yes"  & !is.na(adm_4_weeks)) &
+  #               ((adm_12_weeks == "no" & !is.na(adm_12_weeks)) | is.na(adm_12_weeks)),
+  #             "coca", z)
+  # z <- ifelse(is.na(adm_3_mo) &
+  #               (adm_4_weeks == "yes"  & !is.na(adm_4_weeks)) &
+  #               (adm_12_weeks == "yes" & !is.na(adm_12_weeks)), "coca", z)
+  #
+  # z <- ifelse(z != "hoha" & is.na(adm_3_mo) & is.na(adm_4_weeks) &
+  #               is.na(adm_12_weeks), "all_blank", z)
+
   return(z)
 }